Create HTMLcontactform
This is a simple contact form for a website for my students or anyone. See directions below.  You must replace the replaceyouremailhere@web.com with your real email or the from will not work.  Test the contact form.  First time a message will be sent to your email and you will need to confirm or the contact will not work. The other option is to create a new form in Dreamweaver.   Good luck!  http://multimediatoolkit.me/